Fix Archive->compressCmd()
It was written to try and support eventualy addition of Rar compression
but until that's added we're failing to call setCommand(P7zip::EXTRACT_COMMAND)

Since this is only implemented in P7Zip, we replace it with the Archive::pack()
convenience method, and if/when Rar support gets added down the road we can try to
generalize the initilaztion more. Until then, we have precent already in extractCmd().

Fixes #31